clean up javadoc and jaegerspanexporter name
