+++ instrumentation-core/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/exporters/logging/LoggingSpanExporterAutoConfiguration.java
@@ -26,7 +26,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-/** Create LoggingSpanExporter */
+/** Create LoggingSpanExporter bean*/
 @Configuration
 @EnableConfigurationProperties(LoggingSpanExporterProperties.class)
 @AutoConfigureBefore(TracerAutoConfiguration.class)
